bfix: update all check-in/out components correctly
